chore: move isHTMLElement helper to forge-core
