+++ .changeset/fresh-olives-agree.md
@@ -1,0 +1,5 @@
+---
+'@tylertech/forge-core': patch
+---
+
+feat(utils): add `isHTMLElement` helper
